requirements.txt
Django
django-braces
mysqlclient
six
python-dateutil
pandas